requirements.txt
Django==4.2
django-simple-history==3.3.0